Fix compiling error on linux & macos
